+++ db/migrate/20230302151008_change_all_conditions_met_task_name.rb
@@ -1,0 +1,6 @@
+class ChangeAllConditionsMetTaskName < ActiveRecord::Migration[7.0]
+  def change
+    rename_column :conversion_involuntary_task_lists, :conditions_met_emailed, :conditions_met_confirm_all_conditions_met
+    rename_column :conversion_voluntary_task_lists, :conditions_met_emailed, :conditions_met_confirm_all_conditions_met
+  end
+end