+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   changes to improve words and handle faith schools/diocese
 - the width of the application has been increased from 960px to 1024px when
   viewed on desktop class device.
+- Amended the Confirm all conditions met task content
 
 ### Added